+++ src/janala/interpreters/SymbolicOrConstraint.java
@@ -80,7 +80,7 @@ public class SymbolicOrConstraint extends Constraint {
     }
 
     @Override
-    public Constraint substitute(Map<String, Integer> assignments) {
+    public Constraint substitute(Map<String, Long> assignments) {
         LinkedList<Constraint> tmp = new LinkedList<Constraint>();
         Constraint c2;
         for(Constraint c: constraints) {
